pyproject.toml
[project]
name = "india-branch-data-pipeline"
version = "1.0.0"
description = "90+ web scrapers for Indian bank/NBFC branch data with a complete geocoding and validation pipeline"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "playwright>=1.40.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "googlemaps>=4.10.0",
    "thefuzz>=0.20.0",
    "python-Levenshtein>=0.23.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
